implement db_replay dbapi for data cache server

+++ Development/Src/DataCacheServer/DataCacheDBApiProxy.cpp
@@ -11,8 +11,11 @@ namespace Zion
 		class CProxyDBApi : public IDBApi
 		{
 		public:
+			CProxyDBApi(FILE* fp);
 			CProxyDBApi(IDBApi* db);
 			virtual ~CProxyDBApi();
+
+			bool Read(const char* method, JsonValue& json);
 
 			virtual bool Check();
 			virtual bool Connect();
@@ -32,6 +35,7 @@ namespace Zion
 			virtual bool LockTask(_U32 avatar_id, _U32 task_id);
 			virtual bool DeleteTask(_U32 avatar_id, _U32 task_id);
 
+			FILE* m_fp;
 			IDBApi* m_db;
 			JsonValue* m_json;
 			void* m_userptr;
@@ -43,34 +47,101 @@ namespace Zion
 			return ZION_NEW CProxyDBApi(db);
 		}
 
+		IDBApi* CreateProxyDatabase(const char* filename)
+		{
+			return NULL;
+		}
+
+		CProxyDBApi::CProxyDBApi(FILE* fp)
+		{
+			m_fp = fp;
+			m_db = NULL;
+		}
+
 		CProxyDBApi::CProxyDBApi(IDBApi* db)
 		{
+			m_fp = NULL;
 			m_db = db;
 		}
 
 		CProxyDBApi::~CProxyDBApi()
 		{
-			ZION_DELETE m_db;
-			m_db = NULL;
+			if(m_fp)
+			{
+				fclose(m_fp);
+				m_fp = NULL;
+			}
+			if(m_db)
+			{
+				ZION_DELETE m_db;
+				m_db = NULL;
+			}
+		}
+
+		bool CProxyDBApi::Read(const char* method, JsonValue& json)
+		{
+			static char buf[1000000];
+			if(!fgets(buf, sizeof(buf), m_fp)) return false;
+			char* pos = strchr(buf, ' ');
+			if(!pos) return false;
+			if((size_t)(pos-buf)!=strlen(method) || memcmp(method, buf, pos-buf)!=0) return false;
+			if(!json.Parse(pos+1, pos+strlen(pos+1))) return false;
+			return true;
 		}
 
 		bool CProxyDBApi::Check()
 		{
-			return true;
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("Check", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
+			bool ret = m_db->BeginTranscation();
+			JsonValue json(JsonValue::TYPE_ARRAY);
+			json.Append(JsonValue(ret));
+			WriteDBLog("Check", json);
+			return ret;
 		}
 
 		bool CProxyDBApi::Connect()
 		{
-			return true;
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("Connect", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
+			bool ret = m_db->BeginTranscation();
+			JsonValue json(JsonValue::TYPE_ARRAY);
+			json.Append(JsonValue(ret));
+			WriteDBLog("Connect", json);
+			return ret;
 		}
 
 		bool CProxyDBApi::Disconnect()
 		{
-			return true;
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("Disconnect", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
+			bool ret = m_db->BeginTranscation();
+			JsonValue json(JsonValue::TYPE_ARRAY);
+			json.Append(JsonValue(ret));
+			WriteDBLog("Disconnect", json);
+			return ret;
 		}
 
 		bool CProxyDBApi::BeginTranscation()
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("BeginTranscation", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->BeginTranscation();
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -80,6 +151,12 @@ namespace Zion
 
 		bool CProxyDBApi::RollbackTransaction()
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("RollbackTransaction", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->RollbackTransaction();
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -89,6 +166,12 @@ namespace Zion
 
 		bool CProxyDBApi::CommitTransaction()
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("CommitTransaction", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->CommitTransaction();
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -98,6 +181,12 @@ namespace Zion
 
 		_U32 CProxyDBApi::CreateAvatar(_U32 user_id, _U32 server_id, const char* avatar_name, const char* avatar_desc)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("CreateAvatar", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsU32();
+			}
 			_U32 ret = m_db->CreateAvatar(user_id, server_id, avatar_name, avatar_desc);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -107,6 +196,12 @@ namespace Zion
 
 		bool CProxyDBApi::DeleteAvatar(_U32 avatar_id)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("DeleteAvatar", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->DeleteAvatar(avatar_id);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -132,6 +227,20 @@ namespace Zion
 
 		bool CProxyDBApi::LoadAvatar(_U32 avatar_id, bool (*callback)(void*, const A_UUID&, const char*, const char*), void* userptr)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("LoadAvatar", json)) ZION_FATAL("failed to read db log file\n");
+				const JsonValue& objs = json.Get((_U32)0);
+				for(_U32 i=0; i<objs.GetSize(); i++)
+				{
+					const JsonValue& obj = objs.Get(i);
+					A_UUID _uuid;
+					AUuidFromString(obj.Get((_U32)0).AsCSTR(), _uuid);
+					callback(userptr, _uuid, obj.Get((_U32)1).AsCSTR(), obj.Get((_U32)2).AsCSTR());
+				}
+				return json.Get((_U32)1).AsBool();
+			}
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(JsonValue::TYPE_ARRAY));
 			m_userptr = userptr;
@@ -145,6 +254,12 @@ namespace Zion
 
 		bool CProxyDBApi::InsertAvatarObject(_U32 avatar_id, const A_UUID& _uuid, const char* type, const char* data)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("InsertAvatarObject", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->DeleteAvatar(avatar_id);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -154,6 +269,12 @@ namespace Zion
 
 		bool CProxyDBApi::UpdateAvatarObject(_U32 avatar_id, const A_UUID& _uuid, const char* data)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("UpdateAvatarObject", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->DeleteAvatar(avatar_id);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -163,6 +284,12 @@ namespace Zion
 
 		bool CProxyDBApi::DeleteAvatarObject(_U32 avatar_id, const A_UUID* _uuids, _U32 count)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("DeleteAvatarObject", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->DeleteAvatar(avatar_id);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -172,6 +299,16 @@ namespace Zion
 
 		bool CProxyDBApi::QueryAvatarObject(_U32 avatar_id, const A_UUID& _uuid, bool (*callback)(void*, const A_UUID&, const char*, const char*), void* userptr)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("LoadAvatar", json)) ZION_FATAL("failed to read db log file\n");
+				const JsonValue& obj = json.Get((_U32)0);
+				A_UUID _uuid;
+				AUuidFromString(obj.Get((_U32)0).AsCSTR(), _uuid);
+				callback(userptr, _uuid, obj.Get((_U32)1).AsCSTR(), obj.Get((_U32)2).AsCSTR());
+				return json.Get((_U32)1).AsBool();
+			}
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(JsonValue::TYPE_ARRAY));
 			m_userptr = userptr;
@@ -185,6 +322,12 @@ namespace Zion
 
 		bool CProxyDBApi::LockTask(_U32 avatar_id, _U32 task_id)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("LockTask", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->LockTask(avatar_id, task_id);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
@@ -194,6 +337,12 @@ namespace Zion
 
 		bool CProxyDBApi::DeleteTask(_U32 avatar_id, _U32 task_id)
 		{
+			if(m_fp)
+			{
+				JsonValue json;
+				if(!Read("DeleteTask", json)) ZION_FATAL("failed to read db log file\n");
+				return json.Get((_U32)0).AsBool();
+			}
 			bool ret = m_db->DeleteTask(avatar_id, task_id);
 			JsonValue json(JsonValue::TYPE_ARRAY);
 			json.Append(JsonValue(ret));
